microservices/websocket-service/requirements.txt
Flask
Flask-SocketIO
gunicorn
eventlet
pika
python-dotenv
kombu
